loader_utils: move get_new_subject_split from main
